+++ src/hooks/__tests__/Recoil_useRecoilValueLoadable-test.js
@@ -70,9 +70,9 @@ testRecoil('useRecoilValueLoadable - loadable with error', async () => {
     promise = expect(loadable.toPromise()).rejects.toBeInstanceOf(Error);
     expect(loadable.valueMaybe()).toBe(undefined);
     expect(() => loadable.valueOrThrow()).toThrow(Error);
-    expect((loadable.errorMaybe() ?? {}).toString()).toContain('ERROR');
+    expect(String(loadable.errorMaybe() ?? {})).toContain('ERROR');
     expect(loadable.errorOrThrow()).toBeInstanceOf(Error);
-    expect(loadable.errorOrThrow().toString()).toContain('ERROR');
+    expect(String(loadable.errorOrThrow())).toContain('ERROR');
     expect(loadable.promiseMaybe()).toBe(undefined);
     expect(() => loadable.promiseOrThrow()).toThrow(Error);
     return 'VALUE';